Fix CI -Werror build; address Copilot thumbnail + annotation feedback
- PdfEditor: use size_t stream offsets in snprintf chains; cast QColor
  components to double for printf; fix ink pressure cube math; rename
  ink-list handle to avoid shadowing; size encodeSignature reserves.
- ML/UI: satisfy -Wconversion/-Wsign-conversion in SamSession, OcrEngine,
  BackgroundRemover, FormOverlay, Sidebar, MainWindow, ThumbnailModel.
- Sidebar: handle all AnnotationType values in the annotations list switch.
- ThumbnailModel: drop redundant paper-white compositing (PdfDocument
  already composites); remove standalone thumbnail unit test.
- test_adapters: assert opaque paper-white corner pixels on PDF thumbnails.

+++ src/ui/Sidebar.cpp
@@ -326,7 +326,7 @@ bool Sidebar::eventFilter(QObject *watched, QEvent *event) {
             const auto selected = m_thumbnails->selectionModel()->selectedIndexes();
             if (!selected.isEmpty()) {
                 std::vector<int> pages;
-                pages.reserve(selected.size());
+                pages.reserve(static_cast<size_t>(selected.size()));
                 for (const QModelIndex &idx : selected) {
                     const int p = m_model->pageForRow(idx.row());
                     if (p >= 0)
